Obsolted ADC by OWM - Per email (#166)
* Obsoleted ADC by OWM - Per email

Email from Martin Sippel - As you maybe know, ARRI Media was sold to an investor and is no part of ARRI anymore. We now belong to the Jake Seal Group and our new facility name will be ORWO Media Services GmbH. But all this will just be official at may 1st. But  here are the details as they are in the form:

* Update facilities.json

Co-authored-by: Steve LLamb <38917682+SteveLLamb@users.noreply.github.com>