Update AE to add contact information. Update facilities.json
Per email: Hi. Sorry, I think the confusion was my fault. I was helping a client register a studio code and noticed a mistake with the address on my facility code while I was there. I can’t remember what exactly the mistake was but that’s why I submitted an update. Also, my main site URL has changed so I wanted to update that too.

I might have hit the opt-out thing by mistake. I’m fine with the address being on there. The important thing is that there is a correct URL and/or email address on there for projectionists to contact me if needed.