Add RHF - Red House Films - Update studios.json
Opt OUT

They asked for an update, but I didn't find the original (if it existed).

I don't think it's Red Bull media. The new one is in Mexico. That code is RBMH and they were updating RHF... not a close fit. I think they just clicked the wrong buttons.